Here’s a pull request with three commits that could be split and pulled separately.

The main point is that I wanted more control over which trees were excluded (I often archive in-file, with the tag, and don’t want those trees to be counted). So I added a custom variable for the excluded tags.

I also needed some more control over how links were counted. Some links are references (https://gitlab.com/egh/zotxt) and could be counted as one word. Some links are in-line annotations (https://github.com/girzel/org-annotate) and there the description (which is the "marked" text) should be counted. Some links should perhaps be ignored.

Apart from this, I did some slight modernizing.

I haven’t really tested if this added complexity in `org-word-count-aux` greatly affects speed. I could put up a test document and do it more thoroughly, but I put up the code as a suggestion for now.

Another thing I haven’t tested enough is whether the change of the basic word-skip regex (from "\\w+\\W*" to "\\w+\\s-*",  "\\s-" means whitespace) is a problem. This was needed to avoid skipping into links (inside the [[]]).