@tesujimath 2018-01-16
I did a speed test on the reasonably large org file that I care to count quickly (file size of 150K or so), and it made almost no difference to the time taken (1.6 vs 1.7 seconds when timed by hand, but that's within measurement error), so I think I don't care about any possible slowdown.

If you could test the basic work-skip regex a bit more and let me know that you're happy it didn't break anything, I'll happily merge this.  (I don't really have a plan for automated testing, sorry.)

Thanks for the contribution.