@andersjohansson 2018-01-18
Great!

I’ll have a look at it, test the behaviour with some large files I 
have, and report back. It will probably be in a few days though.