@andersjohansson 2018-01-22
It’s not completely reliable with the changed regex. For example, it jumps right past block beginning lines (at least `#+BEGIN_QUOTE` but not with a space `#+BEGIN_SRC emacs-lisp`) and thus misses detecting blocks correctly. There also seemed to be some case of it skipping into links (which I thought I had avoided) so this needs to be investigated a bit more. Will look back more into it when I have time, but don’t expect this to be ready for merge just yet.

(BTW, I think counting quote blocks is reasonable, so I’m also thinking about abstracting the block detection a bit more to enable this.)