@andersjohansson 2018-01-22
"When I have time" ... Well, I turned the logic around, kept the old word-skipping regex and instead did a save-excursion + go back two chars for checking whether we're inside a link.

Apart from that I fixed a bug in the similar check for latex macros (wrong direction of `<`!) and added a custom variable for blocks which should actually be counted.

It all seems to work fine and as expected for me now.