@andersjohansson 2018-01-22
I suppose it would perhaps cleaner to smash these commits together, since they contain some back and forth changes now.