@tesujimath 2018-01-23
If you squash those commits together, then let me know when you're happy for it to be merged.  Cleaner is better.