@andersjohansson 2018-01-24
There you go! 

I did a very quick test on my current 9000 word text using a simple clock timer:

````emacs-lisp
(defmacro measure-time (&rest body)
  "Measure the time it takes to evaluate BODY."
  `(let ((time (current-time)))
     ,@body
     (message "%.06f" (float-time (time-since time)))))

(measure-time
 (dotimes (i 20)
   (org-word-count (point-min) (point-max))))
````
The new function was actually faster on this file, but I suppose that was because it excluded more:
Old: 8584 words in buffer. [20 times], 8.37 s.
New: 7115 words in buffer. [20 times], 4.65 s.

For a slightly larger file with fewer excluded trees, I get this:
Old: 16684 words in buffer. [20 times], 8.75 s
New: 16866 words in buffer. [20 times], 9.61 s 
(most probably more words since QUOTE blocks are included in the new count)

Also tried with `(org-wc-display nil)`:
Old: 16684 words in buffer. [5 times], 9.78 s
New: 16866 words in buffer. [5 times], 10.56 s

Too me these results seem reasonable.If you want to test it too feel free to do so.