@tesujimath 2018-01-25
Thanks.  I'll check this out as soon as I'm able and merge it if it all seems OK to me now.  (May not be immediately, sorry, have a few other things going on.)