@tesujimath 2018-01-29
I just tested this.  Speed is fine.  However, it now counts inside properties drawers, which the previous version didn't.  Can you fix this?