@andersjohansson 2018-01-30
That’s strange! Can’t immediately figure a good reason why it 
should start doing that now. I’ll have a look some time in the 
next few days.