@andersjohansson 2018-02-06
I have tested now and couldn’t construct any straightforward situation where it counted inside properties drawers (stepping through by edebugging `org-word-count-aux` everything seemed to work as expected). Are you sure?

A simple example which I tried, where `org-word-count` returns `8` (and stepping through shows it skipping the properties drawer).

```
A few words

* A headline
:PROPERTIES:
:CATEGORY: test
:END:

Total words should be 8
```