@tesujimath 2018-02-21
@andersjohansson Just got round to looking at this again.  All seems fine.  Merging.